requirements.txt
apache-airflow==3.0.6
dbt-trino==1.9.3
dbt-core==1.10.10
polars>=0.20.0
connectorx>=0.3.0
sqlalchemy-trino>=0.5.0
boto3>=1.34.0